FastAPI2/requirements.txt
opencv-python~=4.11.0.86
torch~=2.7.0
numpy~=2.2.6
pandas~=2.2.3
pillow~=11.2.1
fastapi~=0.115.12
torchvision~=0.22.0
ultralytics~=8.3.144
protobuf~=5.29.4
google-generativeai